requirements.txt
streamlit
python-dotenv
Pillow
transformers
streamlit-extras
langchain
torch
openai
tokenizers
pydantic